Add model store
* Put the model into a dictory, allowing me to easly add more models to
  the system. And easily re-use all of the other code.
