+++ code/mmEngine/value_funtions/__init__.py
@@ -1,3 +1,3 @@
 from .material import ValueFunctionMaterial, MaterialCount
 from .nn_pytorch import TrainPytorchModel, NNPytorchValueFunction
-from .value_function import ValueFunction, value_function_path
+from .value_function import ValueFunction
