+++ code/mmEngine/value_funtions/value_function.py
@@ -5,7 +5,3 @@ from typing import Optional, Protocol, Union
 class ValueFunction(Protocol):
     def __call__(self, board: Board):
         ...
-
-def value_function_path(name: str) -> Path:
-    value_function_path: Path = Path(__file__).parent
-    return value_function_path / name
