+++ code/simulate_serial.py
@@ -2,8 +2,9 @@ from typing import Optional
 import chess
 from mmEngine.agents import Agent, RandomAgent, MinMaxAgent
 from mmEngine.agents.minMaxAlphaBetaSearch import MinMaxAlphaBetaAgent
-from mmEngine.value_funtions import MaterialCount, value_function_path
+from mmEngine.value_funtions import MaterialCount
 from mmEngine.value_funtions.nn_pytorch import NNPytorchValueFunction, load_model
+from mmEngine.models import load_model, model_store
 from dataclasses import dataclass
 from pathlib import Path
 
@@ -72,9 +73,8 @@ def simulate_games(agent1_factory, agent2_factory, num_games: int=10) -> list[Re
 
 
 def main() -> None:
-    torch_function = "nn.torch"
-    model_path = value_function_path(name=torch_function)
-    fac1 = lambda: MinMaxAlphaBetaAgent(evaluation_function=NNPytorchValueFunction(model=load_model(model_path)), depth=3)
+    model_path, new_model = model_store()["BigCNN"]
+    fac1 = lambda: MinMaxAlphaBetaAgent(evaluation_function=NNPytorchValueFunction(model=load_model(model_path, new_model)), depth=3)
     fac2 = lambda: RandomAgent()
     results = simulate_games(fac1, fac2,num_games=4)
